feat: 강의리뷰등록 구현 (등록시 강의 평균평점 update)

+++ server/src/main/java/ynzmz/server/lecture/dto/LectureDto.java
@@ -31,7 +31,7 @@ public class LectureDto {
         private long lectureId;
         private String name;
         private String introduction;
-        private long starPointAverage;
+        private double starPointAverage;
         private List<Tag.Type> tags;
         private long teacherId;
     }
